# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  models.Address:
    properties:
      hostID:
        type: integer
      ip:
        type: string
      mac:
        type: string
      name:
        type: string
      reserved:
        type: boolean
      state:
        description: network, broadcast, online, offline, reserved, free
        type: string
    type: object
...
  models.IPAM:
    properties:
      addresses:
        description: empty when the subnet is larger than the grid limit
        items:
          $ref: '#/definitions/models.Address'
        type: array
      stat:
        $ref: '#/definitions/models.SubnetStat'
    type: object
...
  models.Stat:
    properties:
      known:
        type: integer
      offline:
        type: integer
      online:
        type: integer
      total:
        type: integer
      unknown:
        type: integer
    type: object
...
  models.Subnet:
    properties:
      cidr:
        type: string
      id:
        type: integer
      iface:
        type: string
      method:
        description: '"arp" or "none"'
        type: string
      name:
        type: string
      reserved:
        description: comma separated IPs and ranges (a-b)
        type: string
      skipPorts:
        description: leave hosts here out of scheduled port scans
        type: boolean
    type: object
  models.SubnetStat:
    properties:
      free:
        type: integer
      nextFree:
        type: string
      offline:
        description: seen before, not online now
        type: integer
      online:
        type: integer
      reserved:
        description: reserved and not in use
        type: integer
      subnet:
        $ref: '#/definitions/models.Subnet'
      total:
        description: usable addresses
        type: integer
      utilization:
        description: percent of usable addresses not free
        type: integer
    type: object
...
info:
  contact:
    url: https://github.com/vipthomps/Gimle
  description: 'Home lab dashboard: network discovery, a map of hosts, bookmarks,
    views and docs'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gimlé API
  version: 1.0.0
paths:
...
  /status/{iface}:
    get:
      description: Retrieve summary statistics of hosts, optionally filtered by interface
      parameters:
      - description: Interface name (omit for all interfaces)
        in: path
        name: iface
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stat'
      summary: Get network status
      tags:
      - system
  /subnets:
    get:
      description: All configured subnets with their address usage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SubnetStat'
            type: array
      summary: List subnets
      tags:
      - subnets
    post:
      consumes:
      - application/json
      description: ID 0 creates a new subnet. Method is "arp" (scan with arp-scan)
        or "none" (track addresses only)
      parameters:
      - description: Subnet
        in: body
        name: subnet
        required: true
        schema:
          $ref: '#/definitions/models.Subnet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subnet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update a subnet
      tags:
      - subnets
  /subnets/{id}:
    delete:
      parameters:
      - description: Subnet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a subnet
      tags:
      - subnets
  /subnets/{id}/ipam:
    get:
      description: Usage and the state of every address in a subnet (addresses omitted
        above 4096)
      parameters:
      - description: Subnet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IPAM'
      summary: Subnet address map
      tags:
      - subnets
...
swagger: "2.0"